+++ app/router.js
@@ -40,6 +40,7 @@ Router.map(function () {
   this.route('mandatarissen', function () {
     this.route('search');
     this.route('persoon', { path: '/:id/persoon' }, function () {
+      this.route('mandaten');
       this.route('mandataris', { path: '/:mandataris_id/mandataris' });
     });
   });
